revise error and various comments
