+++ lib/rbi/validators/duplicates.rb
@@ -26,10 +26,11 @@ module RBI
 
         index.each do |_name, nodes|
           next if nodes.size <= 1
-          next unless nodes.first.is_a?(Method)
 
-          method = T.cast(nodes.first, Method)
-          err = Error.new("Duplicate definitions for `#{method.name}`")
+          first_node = nodes.first
+          next unless first_node.is_a?(Method)
+
+          err = Validators::Error.new("Duplicate definitions for `#{first_node.name}`")
           nodes.each do |node|
             next unless node.is_a?(Method)
 
@@ -41,45 +42,6 @@ module RBI
         end
 
         true
-      end
-
-      class Error < RBI::Error
-        extend T::Sig
-
-        sig { returns(String) }
-        attr_reader :message
-
-        sig { returns(T::Array[Section]) }
-        attr_reader :sections
-
-        sig { params(message: String).void }
-        def initialize(message)
-          super()
-          @message = message
-          @sections = T.let([], T::Array[Section])
-        end
-
-        sig { params(section: Section).void }
-        def <<(section)
-          @sections << section
-        end
-
-        sig { params(loc: T.nilable(Loc)).void }
-        def add_section(loc: nil)
-          self << Section.new(loc: loc)
-        end
-
-        class Section
-          extend T::Sig
-
-          sig { returns(T.nilable(Loc)) }
-          attr_reader :loc
-
-          sig { params(loc: T.nilable(Loc)).void }
-          def initialize(loc: nil)
-            @loc = loc
-          end
-        end
       end
     end
   end
